refactor: 함수 분리 및 상수 파일 관리
