+++ src/js/constants/constants.js
@@ -1,0 +1,4 @@
+export const WEEK = ['일요일', '월요일', '화요일', '수요일', '목요일', '금요일', '토요일'];
+
+export const URL = 'http://localhost:3001';
+
